Revert "sbt DevOops v0.2.0"
