+++ .build-scripts/build-packages.sh
@@ -9,13 +9,13 @@ sbt writeVersion
 
 echo ""
 echo "======================================================"
-echo "Run: sbt -d -J-Xmx2048m clean ^packageBin ^packageSrc ^packageDoc "
+echo "Run: sbt clean ^packageBin ^packageSrc ^packageDoc "
 echo "------------------------------------------------------"
 if sbt clean ^packageBin ^packageSrc ^packageDoc ; then
-  echo "Done: sbt -d -J-Xmx2048m clean ^packageBin ^packageSrc ^packageDoc "
+  echo "Done: sbt clean ^packageBin ^packageSrc ^packageDoc "
   echo "======================================================"
 else
-  echo "Failed: sbt -d -J-Xmx2048m clean ^packageBin ^packageSrc ^packageDoc " 1>&2
+  echo "Failed: sbt clean ^packageBin ^packageSrc ^packageDoc " 1>&2
   echo "======================================================"
   exit 1
 fi
