+++ .build-scripts/release-files-copy.sh
@@ -1,7 +1,5 @@
 #!/bin/bash -e
 
-alias sbt='sbt -d -J-Xmx2048m'
-
 if [ ! -n "$PROJECT_BUILD_NAME" ]
   then
   echo "NO PROJECT_BUILD_NAME is found so quit!" 1>&2
@@ -12,12 +10,12 @@ echo "======================================================"
 echo "Release to Bintray"
 echo "======================================================"
 echo ""
-echo "Run: sbt -d -J-Xmx2048m ^publish"
+echo "Run: sbt ^publish"
 echo "------------------------------------------------------"
-if sbt -d -J-Xmx2048m ^publish ; then
-  echo "Done: sbt -d -J-Xmx2048m ^publish"
+if sbt ^publish ; then
+  echo "Done: sbt ^publish"
 else
-  echo "Failed: sbt -d -J-Xmx2048m ^publish" 1>&2
+  echo "Failed: sbt ^publish" 1>&2
   exit 1
 fi
 echo "======================================================"
@@ -26,8 +24,8 @@ echo "======================================================"
 echo "======================================================"
 echo "Copy packaged files to deploy"
 echo "======================================================"
-echo "ls -l target/scala-*/sbt-*/*.jar"
-ls -l target/scala-*/sbt-*/*.jar
+echo "ls -l target/scala-*/*.jar"
+ls -l target/scala-*/*.jar
 echo ""
 echo "======================================================"
 if [ -d "target/ci" ]; then
@@ -44,8 +42,8 @@ echo "ls -l target/ci/$PROJECT_BUILD_NAME"
 ls -l "target/ci/$PROJECT_BUILD_NAME"
 
 echo "------------------------------------------------------"
-echo "cp target/scala-*/sbt-*/*.jar target/ci/$PROJECT_BUILD_NAME/"
-cp target/scala-*/sbt-*/*.jar "target/ci/$PROJECT_BUILD_NAME/"
+echo "cp target/scala-*/*.jar target/ci/$PROJECT_BUILD_NAME/"
+cp target/scala-*/*.jar "target/ci/$PROJECT_BUILD_NAME/"
 echo "------------------------------------------------------"
 echo "ls -lR target/ci/$PROJECT_BUILD_NAME/"
 ls -lR "target/ci/$PROJECT_BUILD_NAME"
