+++ build.sbt
@@ -8,7 +8,7 @@ lazy val writeVersion = inputKey[Unit]("Write Version in File'")
 
 lazy val root = (project in file("."))
   .settings(
-    organization := "kevinlee"
+    organization := "io.kevinlee"
   , name         := "sbt-devoops"
   , scalaVersion := ProjectScalaVersion
   , version      := ProjectVersion
@@ -24,7 +24,8 @@ lazy val root = (project in file("."))
       ("org.scala-sbt" % "compiler-interface" % sv % "component").sources
     }
   , crossSbtVersions := CrossSbtVersions
-  , scalacOptions ++= crossVersionProps(commonScalacOptions, scalaVersion.value) {
+  , scalacOptions ++=
+      crossVersionProps(commonScalacOptions, scalaVersion.value) {
         case Some((2, 12)) =>
           Seq("-Ywarn-unused-import", "-Ywarn-numeric-widen")
         case Some((2, 11)) =>
@@ -32,22 +33,10 @@ lazy val root = (project in file("."))
         case _ =>
           Nil
       }
-  , scalacOptions in (Compile, console) := scalacOptions.value diff List("-Ywarn-unused-import", "-Xfatal-warnings")
   , wartremoverErrors in (Compile, compile) ++= commonWarts
   , wartremoverErrors in (Test, compile) ++= commonWarts
   , resolvers += Deps.hedgehogRepo
-  , libraryDependencies ++=
-      crossVersionProps(
-          Seq(
-            Deps.commonsIo, Deps.githubApi
-          ) ++ Deps.hedgehogLibs
-        , scalaVersion.value
-      ) {
-        case Some((2, 12)) =>
-          Deps.javaxActivation212
-        case Some((2, 10)) =>
-          Deps.javaxActivation210
-      }
+  , libraryDependencies ++= Seq(Deps.commonsIo) ++ Deps.hedgehogLibs
   , testFrameworks ++= Seq(TestFramework("hedgehog.sbt.Framework"))
 
 //  , addSbtPlugin(Deps.wartRemover)
@@ -60,9 +49,9 @@ lazy val root = (project in file("."))
   , bintrayVcsUrl := Some("""git@github.com:Kevin-Lee/sbt-devoops.git""")
   , bintrayRepository := "sbt-plugins"
 
-  , initialCommands in console := """import kevinlee.sbt._"""
+  , initialCommands in console := """import io.kevinlee.sbt._"""
 
-  , writeVersion := versionWriter(Def.spaceDelimited("filename").parsed)(ProjectVersion)
+  , writeVersion := versionWriter(() => Def.spaceDelimited("filename").parsed)(ProjectVersion)
 
   , coverageHighlighting := (CrossVersion.partialVersion(scalaVersion.value) match {
       case Some((2, 10)) =>
