--- changelogs/0.2.0.md
@@ -1,7 +1,0 @@
-## [0.2.0](https://github.com/Kevin-Lee/sbt-devoops/issues?utf8=✓&q=is%3Aissue+is%3Aclosed+milestone%3A"0.2.0") - 2019-04-07 ([with PR](https://github.com/Kevin-Lee/sbt-devoops/issues?utf8=✓&q=is%3Aclosed+milestone%3A"0.2.0"))
-
-### Done
-* Support GitHub Release (#22)
-* Tag when merge into release branch happens (#23)
-* Copy packaged files (#24)
-* Push tag for release (#32)
