+++ project/ProjectInfo.scala
@@ -1,4 +1,4 @@
-import wartremover.{Wart, Warts}
+import wartremover.WartRemover.autoImport.{Wart, Warts}
 
 /**
   * @author Kevin Lee
@@ -6,15 +6,15 @@ import wartremover.{Wart, Warts}
   */
 object ProjectInfo {
 
-  val ProjectScalaVersion: String = "2.12.7"
-  val CrossScalaVersions: Seq[String] = Seq("2.10.7", ProjectScalaVersion)
+  val ProjectScalaVersion = "2.12.7"
+  val CrossScalaVersions = Seq("2.10.7", ProjectScalaVersion)
 
-  val CrossSbtVersions: Seq[String] = Seq("0.13.17", "1.2.6")
+  val CrossSbtVersions = Vector("0.13.17", "1.2.6")
 
-  val ProjectVersion: String = "0.2.0-SNAPSHOT"
+  val ProjectVersion = "0.1.0"
 
-  val commonScalacOptions: Seq[String] = Seq(
-      "-deprecation"
+  val commonScalacOptions = Seq(
+    "-deprecation"
     , "-unchecked"
     , "-feature"
     , "-Ywarn-value-discard"
@@ -25,17 +25,9 @@ object ProjectInfo {
     , "-Ywarn-inaccessible"
     , "-Ywarn-nullary-unit"
     , "-Ywarn-nullary-override"
-    , "-language:higherKinds"
     , "-encoding", "UTF-8"
   )
 
-  val commonWarts: Seq[Wart] =
-    Warts.allBut(
-        Wart.DefaultArguments
-      , Wart.Overloading
-      , Wart.Any
-      , Wart.Nothing
-      , Wart.NonUnitStatements
-    )
+  val commonWarts = Warts.allBut(Wart.DefaultArguments, Wart.Overloading, Wart.Any, Wart.Nothing, Wart.NonUnitStatements)
 
 }
