--- src/main/scala/kevinlee/fp/SemiGroup.scala
@@ -1,9 +1,0 @@
-package kevinlee.fp
-
-/**
-  * @author Kevin Lee
-  * @since 2019-03-16
-  */
-trait SemiGroup[A] {
-  def append(a1: A, a2: => A): A
-}
